+++ CHANGELOG.md
@@ -3,6 +3,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## master
+### Added
+- Prune devdependencies ([#32](https://github.com/heroku/nodejs-npm-buildpack/pull/32))
 
 ## 0.2.0 (2020-05-19)
 ### Added